Use lexical closures to fix event listening
Using a closure lets us create a rendering function that still has
access to the transdown namespace. That’s important because we want the
renderer to call the `transdownify` function, and the renderer needs
access to the namespace to do so. By defining a closure (var
transdownNamespace = this), the renderer can still access the transdown
namespace even after it’s been bound as a callback function.

refs #3